提交admin 服务

+++ heima-leadnews-model/src/main/java/com/heima/model/mappers/app/ApUserMapper.java
@@ -4,7 +4,7 @@ import com.heima.model.user.pojos.ApUser;
 import org.apache.ibatis.annotations.Param;
 
 public interface ApUserMapper {
-    ApUser selectById(Integer id);
+    ApUser selectById(Long id);
 
     ApUser selectByApPhone(@Param("phone") String phone);
 }
